Added Validated Input For Rows and Columns
The input has some output that we will remove at a later date. The input is also not fully validated as right now it is only checking to make sure that the input is an integer but later we will want to make sure that the integer is a proper value.